[note] compiled the debug build with the MS & Intel compilers with full error reporting enabled & found lots of stuff exp with the Intel Compiler. Most of the changes here are a result of this.
[general]
removed/commented some more unreachable code.
changed all dcxLoadIcon(....TRUE/FALSE...) to be true/false
changed several places where a var called `flags` was being redefined.
removed the optional args from dcxLoadIcon()
removed some code that compared unsigned values to < 0 or checked that it was >= 0, unsigned values will always be >= 0
changed files to end in a newline (odd i know but Intel complains about it)
removed some unneeded ;
removed some unneeded extern's
moved some externs to include files.
added some pragmas to disable Intel Compiler warnings.
commented out un-used var in mIRC_DrawText()

[dialog]
changed PreMultiply()'s nFactor arg to mask out the unwanted bits & type as BYTE.
changed UpdateVistaStyle() with some BYTE typings of pixel in alpha blending routines.

[directshow]
typdef'd some values to avoid errors.

[listview]
changed all commands that add an item to always define LVIF_IMAGE flag to stop the wrong icon being shown when none is wanted.

[dcxml]
fixed DcxmlParser::xdialogEX() & xdidEX() to use txt.to_chr() instead of just the txt object in the formatting strings.
fixed some var redefinitions in ParseIcons().
removed unneeded constructor from DcxmlParser.

[toolbar]
fixed redefinition of numtok var.

[trackbar]
changed to use the SelectBitmap() macro.

[list]
fixed redefinition of numtok var.
changed DrawDragLine() to take a const arg.
changed getDragListId() to return a unsigned int
changed m_iDragList to be an unsigned int

[tinyxml]
added some pragmas to disable Intel Compiler warnings.


git-svn-id: http://dcx.googlecode.com/svn@999 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/dcxdirectshow.cpp
@@ -1072,7 +1072,7 @@ HRESULT DcxDirectshow::setVolume(const float vol)
 	if (SUCCEEDED(hr)) {
 #pragma warning(push,3)
 #pragma warning(disable:4244)
-		long t = -((10000.0 / 100.0) * (100 - vol));
+		long t = (long)-((10000.0 / 100.0) * (100 - vol));
 #pragma warning(pop)
 		hr = pAudio->put_Volume(t);
 		pAudio->Release();
@@ -1091,7 +1091,7 @@ float DcxDirectshow::getVolume() const
 #pragma warning(push,3)
 #pragma warning(disable:4244)
 		if (SUCCEEDED(hr))
-			vol = 100 - ((abs(t) / 10000.0) * 100.0);
+			vol = (float)(100 - ((abs(t) / 10000.0) * 100.0));
 #pragma warning(pop)
 		pAudio->Release();
 	}
